merges SchoolDelegates into DelegateList.
